+++ byterun/interpreter.c
@@ -1,0 +1,9 @@
+//
+// Created by Kirill Golubev on 08.03.2023.
+//
+
+# include "bytecommon.h"
+
+int main(int argc, char*argv[]){
+    return 0;
+}
