+++ runtime/runtime.c
@@ -51,24 +51,9 @@ void __post_gc_subst () {}
 # endif
 /* end */
 
-# define STRING_TAG  0x00000001
-# define ARRAY_TAG   0x00000003
-# define SEXP_TAG    0x00000005
-# define CLOSURE_TAG 0x00000007 
-# define UNBOXED_TAG 0x00000009 // Not actually a tag; used to return from LkindOf
-
-# define LEN(x) ((x & 0xFFFFFFF8) >> 3)
-# define TAG(x)  (x & 0x00000007)
-
-# define TO_DATA(x) ((data*)((char*)(x)-sizeof(int)))
-# define TO_SEXP(x) ((sexp*)((char*)(x)-2*sizeof(int)))
 # ifdef DEBUG_PRINT // GET_SEXP_TAG is necessary for printing from space
 # define GET_SEXP_TAG(x) (LEN(x))
 #endif
-
-# define UNBOXED(x)  (((int) (x)) &  0x0001)
-# define UNBOX(x)    (((int) (x)) >> 1)
-# define BOX(x)      ((((int) (x)) << 1) | 0x0001)
 
 /* GC extra roots */
 # define MAX_EXTRA_ROOTS_NUMBER 32
@@ -155,16 +140,6 @@ void Lassert (void *f, char *s, ...) {
 # define ASSERT_STRING(memo, x)              \
   do if (!UNBOXED(x) && TAG(TO_DATA(x)->tag) \
 	 != STRING_TAG) failure ("string value expected in %s\n", memo); while (0)
-
-typedef struct {
-  int tag; 
-  char contents[0];
-} data; 
-
-typedef struct {
-  int tag; 
-  data contents; 
-} sexp;
 
 extern void* alloc    (size_t);
 extern void* Bsexp    (int n, ...);
@@ -332,9 +307,7 @@ extern int Llength (void *p) {
   return BOX(LEN(a->tag));
 }
 
-static char* chars = "_abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789'";
 
-extern char* de_hash (int);
 
 extern int LtagHash (char *s) {
   char *p;
@@ -390,64 +363,6 @@ char* de_hash (int n) {
   return ++p;
 }
 
-typedef struct {
-  char *contents;
-  int ptr;
-  int len;  
-} StringBuf;
-
-static StringBuf stringBuf;
-
-# define STRINGBUF_INIT 128
-
-static void createStringBuf () {
-  stringBuf.contents = (char*) malloc (STRINGBUF_INIT);
-  memset(stringBuf.contents, 0, STRINGBUF_INIT);
-  stringBuf.ptr      = 0;
-  stringBuf.len      = STRINGBUF_INIT;
-}
-
-static void deleteStringBuf () {
-  free (stringBuf.contents);
-}
-
-static void extendStringBuf () {
-  int len = stringBuf.len << 1;
-
-  stringBuf.contents = (char*) realloc (stringBuf.contents, len);
-  stringBuf.len      = len;
-}
-
-static void vprintStringBuf (char *fmt, va_list args) {
-  int     written = 0,
-          rest    = 0;
-  char   *buf     = (char*) BOX(NULL);
-  va_list vsnargs;
-  
- again:
-  va_copy (vsnargs, args);
-  
-  buf     = &stringBuf.contents[stringBuf.ptr];
-  rest    = stringBuf.len - stringBuf.ptr;
-
-  written = vsnprintf (buf, rest, fmt, vsnargs);
-
-  va_end(vsnargs);
-  
-  if (written >= rest) {
-    extendStringBuf ();
-    goto again;
-  }
-
-  stringBuf.ptr += written;
-}
-
-static void printStringBuf (char *fmt, ...) {
-  va_list args;
-
-  va_start (args, fmt);
-  vprintStringBuf (fmt, args);
-}
 
 int is_valid_heap_pointer (void *p);
 
@@ -1128,6 +1043,8 @@ extern void* Bsexp (int bn, ...) {
 
   return d->contents;
 }
+
+//BsexpTag?
 
 extern int Btag (void *d, int t, int n) {
   data *r;
